+++ src/main/java/com/blackducksoftware/integration/alert/common/enumeration/AlertEnvironment.java
@@ -24,15 +24,24 @@
 package com.blackducksoftware.integration.alert.common.enumeration;
 
 public enum AlertEnvironment {
+    //    ALERT_CONFIG_HOME("ALERT_CONFIG_HOME"),
+    //    ALERT_TEMPLATES_DIR("ALERT_TEMPLATES_DIR"),
+    //    ALERT_IMAGES_DIR("ALERT_IMAGES_DIR"),
+    //    HUB_ALWAYS_TRUST_SERVER_CERTIFICATE("HUB_ALWAYS_TRUST_SERVER_CERTIFICATE"),
+    //    HUB_PROXY_HOST("HUB_PROXY_HOST"),
+    //    HUB_PROXY_PORT("HUB_PROXY_PORT"),
+    //    HUB_PROXY_USER("HUB_PROXY_USER"),
+    //    HUB_PROXY_PASSWORD("HUB_PROXY_PASSWORD"),
+    //    PUBLIC_HUB_WEBSERVER_HOST("PUBLIC_HUB_WEBSERVER_HOST");
     ALERT_CONFIG_HOME("ALERT_CONFIG_HOME"),
     ALERT_TEMPLATES_DIR("ALERT_TEMPLATES_DIR"),
     ALERT_IMAGES_DIR("ALERT_IMAGES_DIR"),
-    HUB_ALWAYS_TRUST_SERVER_CERTIFICATE("HUB_ALWAYS_TRUST_SERVER_CERTIFICATE"),
-    HUB_PROXY_HOST("HUB_PROXY_HOST"),
-    HUB_PROXY_PORT("HUB_PROXY_PORT"),
-    HUB_PROXY_USER("HUB_PROXY_USER"),
-    HUB_PROXY_PASSWORD("HUB_PROXY_PASSWORD"),
-    PUBLIC_HUB_WEBSERVER_HOST("PUBLIC_HUB_WEBSERVER_HOST");
+    ALERT_TRUST_SERVER_CERTIFICATE("ALERT_TRUST_SERVER_CERTIFICATE"),
+    ALERT_PROXY_HOST("ALERT_PROXY_HOST"),
+    ALERT_PROXY_PORT("ALERT_PROXY_PORT"),
+    ALERT_PROXY_USER("ALERT_PROXY_USER"),
+    ALERT_PROXY_PASSWORD("ALERT_PROXY_PASSWORD"),
+    PUBLIC_BLACKDUCK_WEBSERVER_HOST("PUBLIC_BLACKDUCK_WEBSERVER_HOST");
 
     private final String variableName;
