+++ src/main/java/com/blackducksoftware/integration/alert/provider/blackduck/BlackDuckProperties.java
@@ -23,7 +23,6 @@
  */
 package com.blackducksoftware.integration.alert.provider.blackduck;
 
-import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Optional;
 
@@ -38,7 +37,6 @@ import com.blackducksoftware.integration.alert.common.enumeration.AlertEnvironme
 import com.blackducksoftware.integration.alert.common.exception.AlertException;
 import com.blackducksoftware.integration.alert.database.provider.blackduck.GlobalBlackDuckConfigEntity;
 import com.blackducksoftware.integration.alert.database.provider.blackduck.GlobalBlackDuckRepository;
-import com.blackducksoftware.integration.alert.web.model.AboutModel;
 import com.blackducksoftware.integration.exception.EncryptionException;
 import com.blackducksoftware.integration.hub.configuration.HubServerConfig;
 import com.blackducksoftware.integration.hub.configuration.HubServerConfigBuilder;
@@ -48,12 +46,9 @@ import com.blackducksoftware.integration.log.Slf4jIntLogger;
 import com.blackducksoftware.integration.rest.connection.RestConnection;
 import com.blackducksoftware.integration.rest.connection.UnauthenticatedRestConnectionBuilder;
 import com.blackducksoftware.integration.rest.proxy.ProxyInfoBuilder;
-import com.blackducksoftware.integration.util.ResourceUtil;
-import com.google.gson.Gson;
 
 @Component
 public class BlackDuckProperties {
-    public final static String PRODUCT_VERSION_UNKNOWN = "unknown";
     private final GlobalBlackDuckRepository globalBlackDuckRepository;
 
     @Value("${blackduck.hub.url:}")
@@ -74,60 +69,9 @@ public class BlackDuckProperties {
     @Value("${blackduck.hub.proxy.password:}")
     private String blackDuckProxyPassword;
 
-    // SSL properties
-
-    @Value("${server.port:")
-    private String serverPort;
-
-    @Value("${server.ssl.key-store:}")
-    private String keyStoreFile;
-
-    @Value("${server.ssl.key-store-password:}")
-    private String keyStorePass;
-
-    @Value("${server.ssl.keyStoreType:}")
-    private String keyStoreType;
-
-    @Value("${server.ssl.keyAlias:}")
-    private String keyAlias;
-
-    @Value("${server.ssl.trust-store:}")
-    private String trustStoreFile;
-
-    @Value("${server.ssl.trust-store-password:}")
-    private String trustStorePass;
-
-    @Value("${server.ssl.trustStoreType:}")
-    private String trustStoreType;
-
-    private AboutModel aboutModel;
-
     @Autowired
-    public BlackDuckProperties(final GlobalBlackDuckRepository globalBlackDuckRepository, final Gson gson) {
+    public BlackDuckProperties(final GlobalBlackDuckRepository globalBlackDuckRepository) {
         this.globalBlackDuckRepository = globalBlackDuckRepository;
-        readAboutInformation(gson);
-    }
-
-    public void readAboutInformation(final Gson gson) {
-        try {
-            final String aboutJson = ResourceUtil.getResourceAsString(getClass(), "/about.txt", StandardCharsets.UTF_8.toString());
-            aboutModel = gson.fromJson(aboutJson, AboutModel.class);
-        } catch (final Exception e) {
-            aboutModel = null;
-            throw new RuntimeException(e);
-        }
-    }
-
-    public String getProductVersion() {
-        if (aboutModel != null) {
-            return aboutModel.getVersion();
-        } else {
-            return PRODUCT_VERSION_UNKNOWN;
-        }
-    }
-
-    public Optional<AboutModel> getAboutModel() {
-        return Optional.ofNullable(aboutModel);
     }
 
     public String getEnvironmentVariable(final AlertEnvironment alertEnvironment) {
@@ -389,67 +333,4 @@ public class BlackDuckProperties {
         return 300;
     }
 
-    public String getServerPort() {
-        return serverPort;
-    }
-
-    public void setServerPort(final String serverPort) {
-        this.serverPort = serverPort;
-    }
-
-    public String getKeyStoreFile() {
-        return keyStoreFile;
-    }
-
-    public void setKeyStoreFile(final String keyStoreFile) {
-        this.keyStoreFile = keyStoreFile;
-    }
-
-    public String getKeyStorePass() {
-        return keyStorePass;
-    }
-
-    public void setKeyStorePass(final String keyStorePass) {
-        this.keyStorePass = keyStorePass;
-    }
-
-    public String getKeyStoreType() {
-        return keyStoreType;
-    }
-
-    public void setKeyStoreType(final String keyStoreType) {
-        this.keyStoreType = keyStoreType;
-    }
-
-    public String getKeyAlias() {
-        return keyAlias;
-    }
-
-    public void setKeyAlias(final String keyAlias) {
-        this.keyAlias = keyAlias;
-    }
-
-    public String getTrustStoreFile() {
-        return trustStoreFile;
-    }
-
-    public void setTrustStoreFile(final String trustStoreFile) {
-        this.trustStoreFile = trustStoreFile;
-    }
-
-    public String getTrustStorePass() {
-        return trustStorePass;
-    }
-
-    public void setTrustStorePass(final String trustStorePass) {
-        this.trustStorePass = trustStorePass;
-    }
-
-    public String getTrustStoreType() {
-        return trustStoreType;
-    }
-
-    public void setTrustStoreType(final String trustStoreType) {
-        this.trustStoreType = trustStoreType;
-    }
 }
